change usize to u32

+++ mpc-core/src/block.rs
@@ -59,10 +59,10 @@ impl Block {
     }
 
     #[inline]
-    pub fn hash_point(point: &RistrettoPoint, tweak: usize) -> Self {
+    pub fn hash_point(point: &RistrettoPoint, tweak: u32) -> Self {
         let mut hasher = Sha256::new();
         hasher.update(point.compress().as_bytes());
-        hasher.update((tweak as u32).to_be_bytes());
+        hasher.update(tweak.to_be_bytes());
         let h: [u8; 16] = hasher.finalize()[..16]
             .try_into()
             .expect("Unable to convert hash to block");
